requirements.txt
Flask==2.2.2
numpy==1.24.1
Pillow==7.0.0  
psycopg2-binary==2.9.5
tensorflow==2.11.0
Werkzeug==2.2.2